Village 11 12 2021 (#131)
* rp degree int

* tau and fw models fixed

* satur aver service

* stationary solution

* conditions fixed

* removed unused code ui

* m fixed. many questions

* fix

+++ gui/widgets/dataWidget.cpp
@@ -21,7 +21,7 @@ DataWidget::~DataWidget()
 
 void DataWidget::subscribe()
 {
-    auto success = connect(ui->N, SIGNAL(valueChanged(double)), this, SLOT(rp_values_changed()));
+    auto success = connect(ui->N, SIGNAL(valueChanged(int)), this, SLOT(rp_values_changed()));
     Q_ASSERT(success);
     success = connect(ui->MuWat, SIGNAL(valueChanged(double)), this, SLOT(rp_values_changed()));
     Q_ASSERT(success);
