+++ src/common/services/workRp.hpp
@@ -3,12 +3,12 @@
 
 namespace ble::src::common::services::rp {
 
-double get_kw(double s, double n);
-double get_koil(double s, double n);
-double get_fbl(double s, double n, double kmu);
-double get_sigma(double s, double n, double kmu);
-double get_dfbl(double s, double n, double kmu);
-double get_sigma(double s, double n, double kmu, double kw);
+double get_kw(double s, int n);
+double get_koil(double s, int n);
+double get_fbl(double s, int n, double kmu);
+double get_sigma(double s, int n, double kmu);
+double get_dfbl(double s, int n, double kmu);
+double get_sigma(double s, int n, double kmu, double kw);
 
 } // namespace ble::src
